EnableFeature button works after scene switch
(Also debug is granted the same functionality, but does not disappear.)

Introduced a bug where the featurebar disappears after switching scenes to something and back to candy box before enabling all features. Probably an error or discrepancy in case 3 of featureProgress.

I think it may be getting confused between the featureBar and addFeatureButton somehow :/