Small progress on FeatureBar bug
A FeatureBar is spawned from the prefab when it is deleted. It is happening in case 3, and it seems to be deleting the wrong object *shrug*

A band-aid solution may be the only solution at this point. We now need the prefab to act as a true replacement, being a DONTDESTROYONLOAD object and having the proper elements enabled.

We have been changing the GameObject when we should have been changing the prefab 🤦‍♂️